Broke out the Python filters into their own module.
