+++ dxr/plugins/python/__init__.py
@@ -1,3 +1,4 @@
 """ Python plugin, using the ast and tokenize standard library modules. """
 
-from .indexers import FileToIndex, TypeFilter, FunctionFilter, mappings
+from .indexers import FileToIndex, mappings
+from .filters import TypeFilter, FunctionFilter
