+++ dxr/plugins/python/filters.py
@@ -1,0 +1,17 @@
+from jinja2 import Markup
+
+from dxr.plugins.clang.filters import NameFilterBase
+
+
+class _PyFilter(NameFilterBase):
+    lang = 'py'
+
+
+class TypeFilter(_PyFilter):
+    name = 'type'
+    description = Markup("Class definition: <code>type:Stack</code>")
+
+
+class FunctionFilter(_PyFilter):
+    name = 'function'
+    description = Markup("Function or method definition: <code>function:foo</code>")
